+++ packages/app/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hypershape-ai/app
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - dev-server-utils@0.0.4
+  - @hypershape-ai/api@0.0.2
+
 ## 0.0.2
 
 ### Patch Changes